refactor: move start_screen widget to shared game folder
